developed services

+++ src/main/java/com/example/Sport/dto/TrainingRequestDTO.java
@@ -1,7 +1,9 @@
 package com.example.Sport.dto;
 
 import com.example.Sport.bean.Body;
+import com.example.Sport.bean.TypeTraining;
 import com.example.Sport.bean.User;
+import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -19,9 +21,38 @@ public class TrainingRequestDTO {
 
     private Long userId;
 
-    private int time;
+    private Long time;
 
     private Long typeTrainingId;
 
     private Long body;
 }
+
+//@Table(name = "training")
+//public class Training {
+//
+//    @Id
+//    @GeneratedValue(strategy = GenerationType.IDENTITY)
+//    private Long id;
+//
+//    @Column(name = "title")
+//    private String title;
+//
+//    @Column(name = "description_training")
+//    private String description;
+//
+//    @Column(name = "time")
+//    private int time;
+//
+//    @ManyToOne
+//    @JoinColumn(name = "type_training_id")
+//    private TypeTraining typeTraining;
+//
+//    @ManyToOne
+//    @JoinColumn(name = "customer_id")
+//    private User user;
+//
+//    @Enumerated(EnumType.STRING)
+//    @Column(name = "body")
+////  private String trainingBody;
+//    private Body trainingBody;
